agents/event_classifier_agent/agent.py
from google.adk.agents.llm_agent import Agent

root_agent = Agent(
    model='gemini-2.5-flash',
    name='root_agent',
    description='A helpful assistant for user questions.',
    instruction="""
        You are an Event Classifier. Classify the event into one of the following categories:

        - application: Related to programming languages, frameworks, or software development (e.g., Java, Golang, Python, Node.js).
        - infrastructure: Related to servers, networking, databases, cloud infrastructure, or hardware.
        - platform: Related to platforms or tools that support applications (e.g., Kubernetes, Docker, Kafka, CI/CD, AWS services).
        - security: Related to cybersecurity, vulnerabilities, malware, authentication, or attacks.
        - out_of_scope: If the event does not clearly fit any of the above categories or you are unsure.

        Rules:
        - Only respond with one label exactly: application, infrastructure, platform, security, or out_of_scope.
        - Do not explain or add extra text.
        - If unsure, always choose out_of_scope.
        - Treat every event independently.
    """,
)
